make comments more terse.  refactor configuration into default runner directory.  refactor error handling into simple runner directory
